Shell usability improvements, plus logging
This commit tries to improve the usability of the shell script so that
it can be iteratively invoked from the command line without requiring
manual edits.

It can now be invoked like:

./crawl.sh https://www.consumerfinance.gov
./crawl.sh https://www.consumerfinance.gov/start/at/specific/page/

or, to limit its depth to an arbitrary level:

./crawl.sh -d 4 https://www.consumerfinance.gov

This should be helpful for testing and development, for example when
running against localhost, starting at a different URL, or using
different depths.

In addition, the shell now logs wget output to a wget.log logfile, while
still outputting to console. It also logs rejected URLs to rejected.log
using wget's --rejected.log option. This is helpful to debug why certain
URLs were skipped (unfortunately, this file also logs the 2nd, 3rd, nth
time a URL is visited as a "BLACKLIST" rejection even if it was already
crawled).

I've also increased --limit-rate to 1 MB from 200 KB/sec, and removed
--random-wait, which wasn't doing anything since we aren't specifying
--wait.