Improve summary by diffing all intended files
This change modifies the generate_summary.sh script so that it takes an
argument that can be used to indicate which files to diff. Previously,
only HTML files were being diffed, but due to long URLs, wget may
download files that end in other extensions.

Additionally, this changes the workflow file so that we only define the
domain to crawl in a single place.